Move OfferViewModel class to separate file
