+++ OrangeBricks.Web/Controllers/Offers/ViewModels/OffersOnPropertyViewModel.cs
@@ -12,13 +12,4 @@ namespace OrangeBricks.Web.Controllers.Offers.ViewModels
         public IEnumerable<OfferViewModel> Offers { get; set; }
         public int PropertyId { get; set; }
     }
-
-    public class OfferViewModel
-    {
-        public int Id;
-        public int Amount { get; set; }
-        public DateTime CreatedAt { get; set; }
-        public bool IsPending { get; set; }
-        public string Status { get; set; }
-    }
 }
